[NUI] Add ContentPage class
To support different usages of page, Page class becomes abstract class
and ContentPage is newly added.

Page class supports basic page features as follows.
- Navigator : Navigator which has pushed the page.
- PageAppearing and PageDisappearing events.

ContentPage class supports formatted full screen page features as
follows.
- AppBar
- Content
